+++ index.html
@@ -7,8 +7,8 @@
             <h2>Übungen WebAppLecture</h2>
         </div>
         <div class="centered info">
-            <a href="./uebungen/MyMath/myMath.html"><h2>MyMath</h2></a>
-            <a href="./uebungen/Calculator/calculator.html"><h2>Calculator</h2></a>
+            <a href="./uebungen/01-MyMath/myMath.html"><h2>MyMath</h2></a>
+            <a href="./uebungen/02-Calculator/calculator.html"><h2>Calculator</h2></a>
         </div>
     </body>
 </html>
