+++ uebungen/01-MyMath/MyMath.js
