+++ uebungen/01-MyMath/myMath.html
@@ -1,0 +1,17 @@
+<html>
+    <head>
+        <link rel="stylesheet" href="../../src/css/main.css"/>
+        
+    </head>
+    <body>
+        <div class="centered info">
+            Drücke F12 um die Browser Entwicklungsumgebung zu öffnen!<br><br>
+        </div>
+        <div class="centered info">
+            In der Konsole kannst du dann mit der <b>MyMath</b> Klasse experimentieren.<br>
+            z.B.: 
+            <pre>let mmyNumber = new MyMath(69);</pre>
+        </div>
+    </body>
+    <script src="./MyMath.js" type="module"></script>
+</html>
