--- uebungen/MyMath/myMath.css
@@ -1,5 +1,0 @@
-.info {
-    max-width: 250px;
-    margin: 10px auto;
-    border: 5px solid black;
-}
