--- uebungen/MyMath/myMath.html
@@ -1,17 +1,0 @@
-<html>
-    <head>
-        <link rel="stylesheet" href="./myMath.css"/>
-        
-    </head>
-    <body>
-        <div class="info">
-            Drücke F12 um die Browser Entwicklungsumgebung zu öffnen!<br><br>
-        </div>
-        <div class="info">
-            In der Konsole kannst du dann mit der <b>Calculator</b> Klasse experimentieren.<br>
-            z.B.: 
-            <pre>let myCalculator = new Calculator(69);</pre>
-        </div>
-    </body>
-    <script src="./MyMath.js" type="module"></script>
-</html>
